backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.1
python-dotenv>=1.0.1
openai>=1.30.0
